Fix navigation after onboarding, apply to accounting, extract similar logic
